fix patching of nested structs (#22)
* fix patching of nested structs

when nesting structs, the patches should also use nested patches instead of the original nested structs, otherwise a patch with a nested structure can not be deserialized from json (the original non-patched struct is used to deserialize the json and if fields are missing, it fails)

this pr solves the problem by introducing #[patch_name = <name>] attribute that can be added to fields. The name should point to the patch structure that should be used instead of the original. This attribute then causes 2 things to happen:
1. The new name is used for the type of the field in generated code
2. It generates different methods in the Patch traits based on if the fields are renamed or not

* Update struct-patch-derive/src/lib.rs

---------

Co-authored-by: Antonio Yang <yanganto@gmail.com>